remove redundant toLong from PostsMessagesTable

+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/database/tables/PostsMessagesTable.kt
@@ -35,7 +35,7 @@ object PostsMessagesTable : Table() {
                 PostsMessagesTable.postId.eq(postId)
             }.map {
                 PostMessage(
-                    it[messageId].toLong(),
+                    it[messageId],
                     it[mediaGroupId]
                 )
             }
